Share the nobuild GitHub workflow via dev-tools
Stubbing out workflows with replacement no-op workflows is the
recommended way to skip required workflows for some file types but to
require them for others.

A major downside of sharing only some workflows like this is that file
path inclusions / exclusions need to be synchronized across repositories
which can't be done atomically.

See https://github.com/3rdparty/eventuals/issues/176 and
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks