Update the way of grabing files for code style checks
The basic idea is to let users run this script with 1 argument
which specifies necessary directory to be scanned for files
they wanna check. So here is an example how users can call this
script:
```
	./some/path/check_style_of_all_files.sh some/dir
```
Users even can call this script without any argument passed.
If so, then this scipt will search recursively all files needed
for code style checks in the directory from which this script
has been run. For example, immagine that we have the following
structure:
```
--root:
   |
   |___subdirectory1
            |
            |___subdirectory2
            |        |
            |        |
            |        |___dev-tools
            |	             |
            |                |______check-code-style
            |                                |
            |                                |___________check_style_of_all_files.sh
            |
            |___subdirectory3

```
And immagine now that we run script from the 'subdirectory2'. In this
case the script will find only all files which are located recursively
or not in the 'subdirectory2'.