Add Dockerfile.dazel to repo
We are currently using `dazel` in three places. I believe that we should
be able to use the same `Dockerfile` for all three cases. So to ease
maintainability and streamline development I propose this new
Dockerfile.

The Dockerfile solves some problems we've discussed earlier but that
does not seem to be addressed in the two Dockerfiles used in respect
integration.