Add .style.yapf file for standard YAPF (Python) style
Before this PR we were using custom configuration in
`reboot-dev/respect`'s `.vscode.json` to set the desired YAPF style for
our Python code.

With this PR we create a `.style.yapf` file that matches that style,
and make it available for other repos to symlink. This is a step towards
using the same Python style for all repos.