Fix submodule sync
Before this PR, `submodule-sync.yml` assumed that every PR it creates
will run exactly one status check called "Build and Test", which it will
wait to complete before merging. It would sit there and wait, burning an
action runner's worth of resources, while waiting for this test to
complete.

With our recent changes to `respect`'s workflows that repo no longer has
such a status check, meaning that its submodule syncs were never
completing. Also, our tests run long enough now that the wait time for
them to complete is burning a lot of action runner resources.

This PR removes the requirement that `submodule-sync` wait for a status
check. It will immediately assign the `mergequeue-ready` label and stop,
letting mergequeue take it from there.

The downside to this is that `submodule-sync` will no longer file an
issue when the PR fails to pass its checks. We will have to either
create those issues manually or else find a more efficient and flexible
way to respond to those failures.

TESTED: testing submodule-sync is difficult, since a change needs to be
        checked in before it's picked up by workflows like those in
        `reboot-dev/respect`. I'm moderately confident that this simple
        remove-only change will work, so I propose we merge this PR and
        verify post-merge.