Introduce shared isort config file
`isort` is a tool that helps us sort our Python imports in a consistent
way. By having a shared configuration file we can have our VSCode
auto-format our files as we write them.

A future PR can add pre-submit checks that new Python code is formatted
with `isort`, but we'll first need to bring all repos in line with this
new style.