Add `isort` code style check
In #66 we introduced an `isort` config that repos can use to
automatically sort Python `import`s. This commit adds a check to ensure
that the imports are indeed sorted. This check is active only if `isort`
is activated in the repo, as indicated by an `.isort.cfg` file being
present at the top-level of the repo.