Line length: ignore whole files
Before this commit we couldn't get `check_style.sh` to ignore a whole
file; we got pretty far by ignoring only singular lines!

However, now we need to copy in some files that aren't produced by us,
and that we can't reasonably reformat to fit our style. So we need to
have a way to ignore whole files.

This commit adds a new `check_line_length ignore` option, which will
cause the whole remainder of the file it is found in to be ignored.