Fixes to `submodule-sync.yml` (#80)
Before this PR, the `submodule-sync.yml` workflow was failing with a
permissions error (#1805). This PR fixes that by adding an explicit
`permissions` section to the workflow.

While we're here, this PR also deletes some dead code that used to be
used to file issues when the workflow failed. We haven't been
auto-filing such issues in ~forever (the stuck PR is enough of a
signal), so this code was just dead.