`submodule-sync`: make fewer path assumptions
Before this commit,
`per-submodule-build-matrix.sh` assumed that the
top level of the repo it was syncing was at
`../../../../../`, which happened to be true for
all of our repos. As of
https://github.com/reboot-dev/respect/pull/3786/commits/323ce51a506e99ae5452d6c584475edf4a8620fc
however, that is no longer true.

This commit makes the script look up the top-level
of its containing repo dynamically, with its only
assumption being that the script is being run from
a submodule itself - which still holds.

TESTED: checked that the `submodule-sync` workflow
passes again for the `reboot-dev/respect` repo
where it was previously broken.